% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EfalAnnotation}
\alias{EfalAnnotation}
\title{Genome annotation of Eimeria falciformis}
\format{An object of class \code{GRanges} of length 132317.}
\source{
\url{https://toxodb.org/common/downloads/Current_Release/EfalciformisBayerHaberkorn1970/gff/data/ToxoDB-46_EfalciformisBayerHaberkorn1970.gff}
}
\usage{
EfalAnnotation
}
\description{
A dataset containing the annotation of the genome of Eimeria falciformis
}
\examples{
## to load the genome use

}
\keyword{datasets}
