% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EfalGenome}
\alias{EfalGenome}
\title{Genome sequence of Eimeria falciformis}
\format{An object of class \code{ShortRead} of length 753.}
\source{
\url{https://toxodb.org/common/downloads/Current_Release/EfalciformisBayerHaberkorn1970/fasta/data/ToxoDB-46_EfalciformisBayerHaberkorn1970_Genome.fasta}
}
\usage{
EfalGenome
}
\description{
A dataset containing the sequence of the genome of Eimeria falciformis
}
\examples{
## to load the genome use

}
\keyword{datasets}
